refactor: move tooltip and dialog handles to shared location
Move tooltip and dialog handles from routes/__root to shared handles
file to avoid duplication and improve maintainability
