+++ src/components/buttons/update-all.button.tsx
@@ -3,6 +3,7 @@ import { invoke } from "@tauri-apps/api/core";
 import { useState } from "react";
 import { toast } from "sonner";
 
+import type { OutputMod } from "@/components/pages/install-mods";
 import { Button } from "@/components/ui/button";
 import { useAddModUpdateToInstallation } from "@/hooks/use-add-mod-update-to-installation";
 import { installedModsQueryKey } from "@/hooks/use-installed-mods";
@@ -11,7 +12,6 @@ import {
   type ModUpdatesResponse,
   modUpdatesQueryKey,
 } from "@/hooks/use-mod-updates";
-import type { OutputMod } from "@/routes/install-mods/$id";
 import type { Installation } from "@/stores/installations";
 
 export const UpdateAllButton = ({
