+++ src/components/cards/server.card.tsx
@@ -3,10 +3,10 @@ import { DownloadCloudIcon, Lock, Pencil, Play, Star } from "lucide-react";
 import { Button } from "@/components/ui/button";
 import { Group } from "@/components/ui/group";
 import { TooltipTrigger } from "@/components/ui/tooltip";
+import { rootTooltipHandle } from "@/handles";
 import { useDownloadVersion } from "@/hooks/use-download-version";
 import { useInstalledVersionNames } from "@/hooks/use-installed-versions";
 import { cn } from "@/lib/utils";
-import { rootTooltipHandle } from "@/routes/__root";
 import { findInstallationForServer, useInstallations } from "@/stores/installations";
 import type { Server } from "@/stores/servers";
 import { useSettingsStore } from "@/stores/settings";
