+++ src/components/context-menus/installation.context-menu.tsx
@@ -27,12 +27,12 @@ import {
   ContextMenuTrigger,
 } from "@/components/ui/context-menu";
 import { DialogTrigger } from "@/components/ui/dialog";
+import { rootAlertDialogHandle, rootDialogHandle } from "@/handles";
 import { useDownloadVersion } from "@/hooks/use-download-version";
 import { useInstalledVersionNames } from "@/hooks/use-installed-versions";
 import { usePlayInstallation } from "@/hooks/use-play-installation";
 import { useRevealInFolder } from "@/hooks/use-reveal-in-folder";
 import { cn, exportInstallation } from "@/lib/utils";
-import { rootAlertDialogHandle, rootDialogHandle } from "@/routes/__root";
 import { type Installation, useInstallations } from "@/stores/installations";
 
 export const InstallationContextMenu = ({
