+++ src/components/menus/installation.menu.tsx
@@ -18,12 +18,12 @@ import { ViewLogsDialog } from "@/components/dialogs/viewlogs.dialog";
 import { AlertDialogTrigger } from "@/components/ui/alert-dialog";
 import { DialogTrigger } from "@/components/ui/dialog";
 import { MenuGroup, MenuItem } from "@/components/ui/menu";
+import { rootAlertDialogHandle, rootDialogHandle } from "@/handles";
 import { useDownloadVersion } from "@/hooks/use-download-version";
 import { useInstalledVersionNames } from "@/hooks/use-installed-versions";
 import { usePlayInstallation } from "@/hooks/use-play-installation";
 import { useRevealInFolder } from "@/hooks/use-reveal-in-folder";
 import { cn, exportInstallation } from "@/lib/utils";
-import { rootAlertDialogHandle, rootDialogHandle } from "@/routes/__root";
 import { type Installation, useInstallations } from "@/stores/installations";
 
 export const InstallationMenu = ({ installation }: { installation: Installation }) => {
