+++ src/components/rows/installation.row.tsx
@@ -24,17 +24,17 @@ import { DialogTrigger } from "@/components/ui/dialog";
 import { Group, GroupSeparator } from "@/components/ui/group";
 import { MenuTrigger } from "@/components/ui/menu";
 import { TooltipTrigger } from "@/components/ui/tooltip";
-import { useDownloadVersion } from "@/hooks/use-download-version";
-import { useInstalledVersionNames } from "@/hooks/use-installed-versions";
-import { usePlayInstallation } from "@/hooks/use-play-installation";
-import { useRevealInFolder } from "@/hooks/use-reveal-in-folder";
-import { cn, exportInstallation } from "@/lib/utils";
 import {
   rootAlertDialogHandle,
   rootDialogHandle,
   rootMenuHandle,
   rootTooltipHandle,
-} from "@/routes/__root";
+} from "@/handles";
+import { useDownloadVersion } from "@/hooks/use-download-version";
+import { useInstalledVersionNames } from "@/hooks/use-installed-versions";
+import { usePlayInstallation } from "@/hooks/use-play-installation";
+import { useRevealInFolder } from "@/hooks/use-reveal-in-folder";
+import { cn, exportInstallation } from "@/lib/utils";
 import { type Installation, useInstallations } from "@/stores/installations";
 
 export type InstallationRowProps = {
